projects/omrs-appo-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.9
aiohttp==3.11.0
redis[hiredis]==5.2.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
httpx==0.27.2
fhir.resources==8.0.0
google-generativeai==0.8.3
twilio==9.3.7
python-dotenv==1.0.1
loguru==0.7.2
jinja2==3.1.4
pendulum==3.0.0
tenacity==9.0.0